REFACTOR: centralize MNE and array data extraction
