+++ mne_denoise/asr/adaptive.py
@@ -27,9 +27,10 @@ from typing import Any
 
 import numpy as np
 
+from .._data import extract_data_from_mne
 from .._logging import set_log_level_from_verbose
 from .._spatial import continuous_to_epochs
-from ..utils import extract_data_from_mne, reconstruct_mne_object
+from ..utils import reconstruct_mne_object
 from ._covariance import (
     _adaptive_covariance_sqrt,
 )
