+++ mne_denoise/icanclean/core.py
@@ -60,9 +60,10 @@ from scipy import linalg as la
 from sklearn.base import BaseEstimator, TransformerMixin
 
 from .._cca import canonical_correlation
+from .._data import extract_data_from_mne
 from .._filtering import _filter_channels
 from .._logging import set_log_level_from_verbose
-from ..utils import extract_data_from_mne, reconstruct_mne_object
+from ..utils import reconstruct_mne_object
 
 # Optional MNE support
 try:
