+++ mne_denoise/sound/core.py
@@ -37,10 +37,11 @@ import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
+from .._data import extract_data_from_mne
 from .._leadfield import resolve_leadfield
 from .._logging import logger, set_log_level_from_verbose
 from .._spatial import epochs_to_continuous
-from ..utils import extract_data_from_mne, reconstruct_mne_object
+from ..utils import reconstruct_mne_object
 
 
 def _noise_level(noise_filter: np.ndarray, cov: np.ndarray, n_times: int) -> float:
